Realtime safe buffer reference handling.
Turns out that dropping a shared_ptr is not realtime safe, even if you use
a realtime safe deleter.  Instead, instrusive_ptr is used for buffer
references, so a buffer reference may safely be dropped in the audio
thread (in which case it will be recycled by the BufferFactory).

Faster, cleaner, better.


git-svn-id: http://svn.drobilla.net/lad@2341 a436a847-0d15-0410-975c-d299462d15a1