Add center parameter to FlowCanvas::Canvas::arrange with default value true (eqivalent default behaviour to API pre-r2496).
Add FlowCanvas::Canvas::move_contents_to to allow apps to move canvas contents
to a given location (e.g. the origin or the center), keeping their relative
distances the same.

This commit preserves centering by default with the option to arrange in
the top left, but applications shouldn't assume anything about the origin
of coordinates (i.e. applications should manually scroll to where items
exist when loading a set of items with coordinates, not simply assume
they are centered about a particular location).


git-svn-id: http://svn.drobilla.net/lad@2500 a436a847-0d15-0410-975c-d299462d15a1